pushdown table name filter to system.iceberg_history for catalogs
